{"types":"C","runs":4}

+++ problem15.java
@@ -2,8 +2,8 @@ import acm.program.ConsoleProgram;
 
 public class problem15 extends ConsoleProgram{
     public void run(){
-    	int n = readInt("ENTER THE FIRST NUMBER: " );
-    	int m = readInt("ENTER THE FIRST NUMBER: " );
+    	double n = readInt("ENTER THE FIRST NUMBER: " );
+    	double m = readInt("ENTER THE FIRST NUMBER: " );
     	double x = (n + m) / 2;
     	print("THE ANSWER IS: " + x);
     }
